requirements_new.txt
# 现有依赖
langchain
langchain-openai
langgraph
sounddevice
soundfile
opencv-python
gradio
numpy

# 新增依赖
sentence-transformers==2.2.2
faiss-cpu==1.7.4
beautifulsoup4==4.12.2
requests==2.31.0
scikit-learn==1.3.0